New periodic behavior for T-Res tasks:
Notifications from input sensors are put in a list (integers). The list can be
retrieved by the script which is executed periodically.

New configuration parameter to set the period (when the task is created, query
string of the PUT request)

New API getInputList(), allows to retrieve the list of integers collected
during the pervious period